Day 2

+++ src/day01.zig
@@ -2,13 +2,13 @@ const std = @import("std");
 
 pub fn day01a(allocator: std.mem.Allocator, file: []const u8) ![]const u8 {
     var line_iterator = std.mem.splitScalar(u8, file, '\n');
-    var curr: u16 = try std.fmt.parseInt(u16, line_iterator.first(), 10);
-    var up_counter: u16 = 0;
+    var curr: u64 = try std.fmt.parseInt(u64, line_iterator.first(), 10);
+    var up_counter: u64 = 0;
     while (line_iterator.next()) |line| {
         if (line.len == 0) {
             break;
         }
-        const next: u16 = try std.fmt.parseInt(u16, line, 10);
+        const next: u64 = try std.fmt.parseInt(u64, line, 10);
         if (next > curr) {
             up_counter += 1;
         }
@@ -18,7 +18,7 @@ pub fn day01a(allocator: std.mem.Allocator, file: []const u8) ![]const u8 {
 }
 
 pub fn day01b(allocator: std.mem.Allocator, file: []const u8) ![]const u8 {
-    var nums = std.ArrayList(u16).init(allocator);
+    var nums = std.ArrayList(u64).init(allocator);
     defer nums.deinit();
 
     var line_iterator = std.mem.splitScalar(u8, file, '\n');
@@ -26,21 +26,21 @@ pub fn day01b(allocator: std.mem.Allocator, file: []const u8) ![]const u8 {
         if (line.len == 0) {
             break;
         }
-        const num = try std.fmt.parseInt(u16, line, 10);
+        const num = try std.fmt.parseInt(u64, line, 10);
         try nums.append(num);
     }
 
-    var sums = std.ArrayList(u16).init(allocator);
+    var sums = std.ArrayList(u64).init(allocator);
     defer sums.deinit();
 
     try sums.resize(nums.items.len - 2);
     for (0..(nums.items.len - 2)) |index| {
         const relevant_lines = nums.items[index .. index + 3];
-        var sum: u16 = 0;
+        var sum: u64 = 0;
         for (relevant_lines) |line| sum += line;
         sums.items[index] = sum;
     }
-    var up_counter: u16 = 0;
+    var up_counter: u64 = 0;
     var curr_sum = sums.items[0];
     for (sums.items[1..]) |sum| {
         if (sum > curr_sum) {
